fix: harden issue sync closeout
