+++ scripts/pr/issue-board-transport.mjs
@@ -8,7 +8,12 @@
 
 import { spawn } from "node:child_process";
 
-import { splitRepo, validateOpenPr } from "./issue-board-state.mjs";
+import {
+  labelNames,
+  labelsForState,
+  splitRepo,
+  validateOpenPr,
+} from "./issue-board-state.mjs";
 
 const GH_OUTPUT_MAX_BYTES = 20 * 1024 * 1024;
 // A bounded traversal must fail rather than use an incomplete comment history.
@@ -107,6 +112,27 @@ export function runGh(args, { dryRun = false, mutates = false } = {}) {
       resolve(stdout);
     });
   });
+}
+
+export async function editIssueLabels(options, issue, state) {
+  const transition = labelsForState(state);
+  const existingLabels = labelNames(issue);
+  const addLabels = transition.addLabels.filter(
+    (label) => !existingLabels.has(label),
+  );
+  const removeLabels = transition.removeLabels.filter((label) =>
+    existingLabels.has(label),
+  );
+  if (addLabels.length === 0 && removeLabels.length === 0) return;
+
+  const args = ["issue", "edit", String(issue.number), "-R", options.repo];
+  if (addLabels.length > 0) {
+    args.push("--add-label", addLabels.join(","));
+  }
+  if (removeLabels.length > 0) {
+    args.push("--remove-label", removeLabels.join(","));
+  }
+  await runGh(args, { dryRun: options.dryRun, mutates: true });
 }
 
 export async function ghJson(args, opts = {}) {
